compliance-dashboard/requirements.txt
# Compliance Dashboard Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9

# Environment
python-dotenv==1.0.0

# Data Processing
python-multipart==0.0.6

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Optional: PDF processing (for future sessions)
# pypdf2==3.0.1
# reportlab==4.0.9
